+++ modal-js/src/cls.ts
@@ -4,6 +4,7 @@ import {
   ClassParameterSet,
   ClassParameterSpec,
   ClassParameterValue,
+  FunctionHandleMetadata,
   FunctionOptions,
   FunctionRetryPolicy,
   ParameterType,
@@ -61,23 +62,10 @@ export class ClsService {
         );
       }
 
-      let methodNames: string[];
-      if (serviceFunction.handleMetadata?.methodHandleMetadata) {
-        methodNames = Object.keys(
-          serviceFunction.handleMetadata.methodHandleMetadata,
-        );
-      } else {
-        // Legacy approach not supported
-        throw new Error(
-          "Cls requires Modal deployments using client v0.67 or later.",
-        );
-      }
       return new Cls(
         this.#client,
         serviceFunction.functionId,
-        schema,
-        methodNames,
-        serviceFunction.handleMetadata?.inputPlaneUrl,
+        serviceFunction.handleMetadata!,
         undefined,
       );
     } catch (err) {
@@ -123,26 +111,24 @@ type ServiceOptions = ClsWithOptionsParams & {
 export class Cls {
   #client: ModalClient;
   #serviceFunctionId: string;
-  #schema: ClassParameterSpec[];
-  #methodNames: string[];
-  #inputPlaneUrl?: string;
+  #serviceFunctionMetadata: FunctionHandleMetadata;
   #serviceOptions?: ServiceOptions;
 
   /** @ignore */
   constructor(
     client: ModalClient,
     serviceFunctionId: string,
-    schema: ClassParameterSpec[],
-    methodNames: string[],
-    inputPlaneUrl?: string,
+    serviceFunctionMetadata: FunctionHandleMetadata,
     options?: ServiceOptions,
   ) {
     this.#client = client;
     this.#serviceFunctionId = serviceFunctionId;
-    this.#schema = schema;
-    this.#methodNames = methodNames;
-    this.#inputPlaneUrl = inputPlaneUrl;
+    this.#serviceFunctionMetadata = serviceFunctionMetadata;
     this.#serviceOptions = options;
+  }
+
+  get #schema(): ClassParameterSpec[] {
+    return this.#serviceFunctionMetadata.classParameterInfo?.schema ?? [];
   }
 
   /**
@@ -164,25 +150,26 @@ export class Cls {
     } else {
       functionId = await this.#bindParameters(parameters);
     }
+
     const methods = new Map<string, Function_>();
-    for (const name of this.#methodNames) {
+    for (const [name, methodMetadata] of Object.entries(
+      this.#serviceFunctionMetadata.methodHandleMetadata,
+    )) {
       methods.set(
         name,
-        new Function_(this.#client, functionId, name, this.#inputPlaneUrl),
+        new Function_(this.#client, functionId, name, methodMetadata),
       );
     }
     return new ClsInstance(methods);
   }
 
   /** Override the static Function configuration at runtime. */
-  withOptions(params: ClsWithOptionsParams): Cls {
-    const merged = mergeServiceOptions(this.#serviceOptions, params);
+  withOptions(options: ClsWithOptionsParams): Cls {
+    const merged = mergeServiceOptions(this.#serviceOptions, options);
     return new Cls(
       this.#client,
       this.#serviceFunctionId,
-      this.#schema,
-      this.#methodNames,
-      this.#inputPlaneUrl,
+      this.#serviceFunctionMetadata,
       merged,
     );
   }
@@ -196,9 +183,7 @@ export class Cls {
     return new Cls(
       this.#client,
       this.#serviceFunctionId,
-      this.#schema,
-      this.#methodNames,
-      this.#inputPlaneUrl,
+      this.#serviceFunctionMetadata,
       merged,
     );
   }
@@ -212,9 +197,7 @@ export class Cls {
     return new Cls(
       this.#client,
       this.#serviceFunctionId,
-      this.#schema,
-      this.#methodNames,
-      this.#inputPlaneUrl,
+      this.#serviceFunctionMetadata,
       merged,
     );
   }