+++ test-support/setup.sh
@@ -37,5 +37,10 @@ modal secret create --force libmodal-anthropic-secret \
   ANTHROPIC_API_KEY="$anthropic_api_key_secret" \
   >/dev/null
 
+# deploy an app using an older version of Modal that's unsupported by libmodal
+uv venv modal_1_1
+uv pip install --python modal_1_1 "modal<1.2"
+modal_1_1/bin/modal deploy test_support_1_1.py
+
 echo
 echo "NOTE! The tests also require a Proxy named 'libmodal-test-proxy', which cannot be created programmatically and must be created using the dashboard: https://modal.com/settings/modal-labs/proxy"